reorganize folder
sounds folder is now reorganized into categories. A naming convention has been adopted which identifies the activity taking place. If no activity can be matched, a misc. with the activity name is in the activity folder for which the trigger is to fire. This naming convention is more efficient when coding the play sound triggers and other actions which will take place in future.